Implemented Edit API method partly

+++ www/api.php
@@ -20,7 +20,7 @@ try
 	$user_name = NULL;
 	$user_perms = array();
 
-	if (count($xml->Login->children()) == 2)
+	if (isset($xml->Login->Username) && isset($xml->Login->Password))
 	{
 		$stmt = $db->prepare("SELECT * FROM users WHERE username = ?");
 		$stmt->bind_param("s", $xml->Login->Username);
@@ -79,15 +79,20 @@ try
 			if (in_array("p_delete", $user_perms))
 			{
 				$post_id = (int)$xml->ID;
-				$result = $db->query("SELECT mime FROM posts WHERE id = " . $post_id);
+				$result = $db->query("SELECT user_id, private, mime FROM posts WHERE id = " . $post_id);
 				if ($result->num_rows == 1)
 				{
-					$mime = $result->fetch_row()[0];
-					$db->query("DELETE FROM post_tags WHERE post_id = " . $post_id);
-					$db->query("DELETE FROM posts WHERE id = " . $post_id);
-					unlink($image_dir . "image" . $post_id . $mime_types[$mime]);
-					unlink($thumb_dir . "thumb" . $post_id . ".jpg");
-					api_result_noerror();
+					$row = $result->fetch_assoc();
+					if ($row["private"] == 0 || $row["user_id"] == $user_id || in_array("p_admin", $user_perms))
+					{
+						$mime = $row["mime"];
+						$db->query("DELETE FROM post_tags WHERE post_id = " . $post_id);
+						$db->query("DELETE FROM posts WHERE id = " . $post_id);
+						unlink($image_dir . "image" . $post_id . $mime_types[$mime]);
+						unlink($thumb_dir . "thumb" . $post_id . ".jpg");
+						api_result_noerror();
+					}
+					else throw new Exception("Access denied");
 				}
 				else throw new Exception("Post not found");
 			}
@@ -105,7 +110,60 @@ try
 				if ($result->fetch_row()[0] == 1)
 					echo "\t<Bool>1</Bool>\n";
 				else echo "\t<Bool>0</Bool>\n";
+			} break;
+
+		case "Edit":
+			if (in_array("p_edit", $user_perms))
+			{
+				$post_id = (int)$xml->ID;
+				$result = $db->query("SELECT user_id, private FROM posts WHERE id = " . $post_id);
+				if ($result->num_rows == 1)
+				{
+					$row = $result->fetch_assoc();
+					if ($row["private"] == 0 || $row["user_id"] == $user_id || in_array("p_admin", $user_perms))
+					{
+						if (isset($xml->Post->Source))
+						{
+							$nsource = (string)$xml->Post->Source;
+							$stmt = $db->prepare("UPDATE posts SET source = ? WHERE id = ?");
+							$stmt->bind_param("si", $nsource, $post_id);
+							$stmt->execute();
+						}
+						if (isset($xml->Post->Info))
+						{
+							$ninfo = (string)$xml->Post->Info;
+							$stmt = $db->prepare("UPDATE posts SET info = ? WHERE id = ?");
+							$stmt->bind_param("si", $ninfo, $post_id);
+							$stmt->execute();
+						}
+						if (isset($xml->Post->Rating))
+						{
+							$nrating = (int)$xml->Post->Rating;
+							$stmt = $db->prepare("UPDATE posts SET rating = ? WHERE id = ?");
+							$stmt->bind_param("ii", $nrating, $post_id);
+							$stmt->execute();
+						}
+						if (isset($xml->Post->Private))
+						{
+							if ($row["user_id"] == $user_id || in_array("p_admin", $user_perms))
+							{
+								$nprivate = ($xml->Post->Private > 0) ? 1 : 0;
+								$stmt = $db->prepare("UPDATE posts SET private = ? WHERE id = ?");
+								$stmt->bind_param("ii", $nprivate, $post_id);
+								$stmt->execute();
+							}
+							else throw new Exception("Only the owner can make the post private");
+						}
+						if (isset($xml->Post->Tags))
+						{
+						}
+						api_result_noerror();
+					}
+					else throw new Exception("Access denied");
+				}
+				else throw new Exception("Post not found");
 			}
+			else throw new Exception("No edit permission");
 			break;
 	}
 }
